# dist/swagger.yaml
...
swagger: '2.0'
info:
  title: SolarREST Server
  description: Needs more Panel
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /panel/postPanel:
    post:
      summary: Neues Solarpanel erstellen
      description: Erstellt ein neues Solarpanel auf dem Server. Das ID Feld des Panel kann leer gelassen werden, da dieses vom Server generiert werden muss. Alle anderen Felder muessen ausgefüllt sein.
      parameters:
        - name: Solarpanel
          in: body
          description: Solarpanel welches erstellt werden soll
          required: true
          schema:
            $ref: '#/definitions/Solarpanel'
      tags:
        - Solarpanel
      responses:
        200:
          description: Das erzeugte Solarpanel mit ID
          schema:
            $ref: '#/definitions/Solarpanel'
  /panel/updatePanel:
    post:
      summary: Update eines bestenden Panels
      description: Updateted das Panel, falls es beispielsweise vom User verschoben wurde.
      parameters:
        - name: Solarpanel
          in: body
          description: Solarpanel welches geupdated werden soll
          required: true
          schema:
           $ref: '#/definitions/Solarpanel'
      tags:
        - Solarpanel
      responses:
        200:
          description: Das erzeugte Solarpanel mit ID
          schema:
            $ref: '#/definitions/Solarpanel'
...
definitions:
...
  LatitudeLongitude:
    type: object
    properties:
      latitude:
        type: number
        format: double
        description: Latitude eines Punktes.
      longitutde:
        type: number
        format: double
        description: Longitude eines Punkes.
  Solarpanel:
    type: object
    properties:
      panel_id:
        type: integer
        description: Eindeutige ID eines Solarpanels.
      laenge:
        type: number
        format: double
        description: Länge eines Solarpanels.
      breite:
        type: number
        format: double
        description: Breite eines Solarpanels.
      neigung:
        type: number
        format: double
        description: Neigung eines Solarpanels.
      ausrichtung:
        type: number
        format: double
        description: Ausrichtung eines Solarpanles.
      rahmenbreite:
        type: number
        format: double
        description: Breite des Rahmens eines Solarpanels.
      the_geom:
        type: array
        items:
          $ref: '#/definitions/LatitudeLongitude'
...